+++ lib/data_classes/match/index.dart
@@ -2,3 +2,4 @@ export "./combined_match.dart";
 export "./match.dart";
 export "./match_filter.dart";
 export "./match_sort.dart";
+export './match_filter_value.dart';
