+++ lib/screens/player_detail/player_detail_filter_tab.dart
@@ -14,6 +14,9 @@ class PlayerDetailFilterTab extends HookConsumerWidget {
   Widget build(BuildContext context, WidgetRef ref) {
     // Providers
     final matchesProvider = ref.read(providers.matches);
+    final isGuest = ref.watch(
+      providers.auth.select((_) => _.isGuest),
+    );
     final selectedFilter = ref.watch(
       providers.matches.select((_) => _.selectedFilter),
     );
@@ -64,7 +67,7 @@ class PlayerDetailFilterTab extends HookConsumerWidget {
       padding: const EdgeInsets.symmetric(horizontal: 10),
       child: ListView(
         physics: const ClampingScrollPhysics(),
-        children: data_classes.MatchFilter.filterNames.map(
+        children: data_classes.MatchFilter.filterNames(isGuest).map(
           (filterName) {
             final isFilterOpened = openedFilterName.value == filterName;
             final isFilterNameSelected = selectedFilter.name == filterName;
